fix nodes module
