+++ pyobis/nodes/nodes.py
@@ -3,10 +3,10 @@
 """
 from urllib.parse import urlencode
 
-from ..obisutils import obis_baseurl, obis_GET
+from ..obisutils import OBISQueryResult, obis_baseurl, obis_GET
 
 
-class OBISQueryResult:
+class NodesQuery(OBISQueryResult):
     def __init__(self):
         """
         An OBISQueryResult object for fetching occurrence records.
@@ -26,11 +26,11 @@ class OBISQueryResult:
             nodes = OQR()
             nodes.search(id="4bf79a01-65a9-4db6-b37b-18434f26ddfc")
         """
-        self.url = obis_baseurl + "node/" + id
+        OBISQueryResult.url = obis_baseurl + "node/" + id
         self.mapper = True
-        self.args = {}
+        OBISQueryResult.args = {}
         self.nodeid = id  # necessary to get mapper url
-        out = obis_GET(self.url, self.args, "application/json; charset=utf-8", **kwargs)
+        out = obis_GET(OBISQueryResult.url, OBISQueryResult.args, "application/json; charset=utf-8", **kwargs)
 
         return out
 
@@ -48,31 +48,11 @@ class OBISQueryResult:
             nodes = OQR()
             nodes.activities(id="4bf79a01-65a9-4db6-b37b-18434f26ddfc")
         """
-        self.url = obis_baseurl + "node/" + id + "/activities"
-        self.args = {}
+        OBISQueryResult.url = obis_baseurl + "node/" + id + "/activities"
+        OBISQueryResult.args = {}
         self.mapper = False
-        out = obis_GET(self.url, self.args, "application/json; charset=utf-8", **kwargs)
+        out = obis_GET(OBISQueryResult.url, OBISQueryResult.args, "application/json; charset=utf-8", **kwargs)
         return out
-
-    def get_search_url(self):
-        """
-        Get the corresponding API URL for the query.
-
-        :return: OBIS API URL for the corresponding query
-
-        Usage::
-
-            from pyobis.checklist import OBISQueryresult as OQR
-            query = OQR()
-            data = query.list(scientificname="Mola mola")
-            api_url = query.get_search_url()
-            print(api_url)
-        """
-        return (
-            self.url
-            + "?"
-            + urlencode({k: v for k, v in self.args.items() if v is not None})
-        )
 
     def get_mapper_url(self):
         """
