+++ pyobis/obisutils.py
@@ -1,6 +1,7 @@
 """
 Utility functions for internal use across various modules.
 """
+from urllib.parse import urlencode
 import requests
 
 obis_baseurl = "https://api.obis.org/v3/"
@@ -13,6 +14,21 @@ class NoResultException(Exception):
 
     pass
 
+class OBISQueryResult:
+    """
+    Return OBIS API URL
+    """    
+    def get_search_url(self):
+        """
+        Get the corresponding API URL for the query.
+
+        :return: OBIS API URL for the corresponding query
+        """
+        return (
+            self.url
+            + "?"
+            + urlencode({k: v for k, v in self.args.items() if v is not None})
+        )
 
 def obis_GET(url, args, ctype, **kwargs):
     """Handles technical details of sending GET request to the API"""
